+++ infrastructure/src/main/java/com/music/s3/S3Storage.java
@@ -5,7 +5,6 @@ import com.amazonaws.AmazonServiceException;
 import com.amazonaws.event.ProgressListener;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.AmazonS3Exception;
-import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.transfer.MultipleFileUpload;
 import com.amazonaws.services.s3.transfer.ObjectMetadataProvider;
@@ -13,10 +12,10 @@ import com.amazonaws.services.s3.transfer.TransferManager;
 import com.music.constatns.AudioQuality;
 import com.music.constatns.FileType;
 import com.music.adaptor.FileStorage;
-import com.music.upload.dto.LyricsUpload;
-import com.music.upload.dto.MusicUpload;
-import com.music.upload.dto.FileKeys;
-import com.music.upload.dto.TrackUpload;
+import com.music.dto.upload.LyricsUpload;
+import com.music.dto.upload.MusicUpload;
+import com.music.dto.upload.FileKeys;
+import com.music.dto.upload.TrackUpload;
 import com.music.util.FileUtils;
 import java.io.File;
 import java.io.IOException;
