+++ music/src/main/java/com/music/dto/upload/FileKeys.java
@@ -1,4 +1,4 @@
-package com.music.upload.dto;
+package com.music.dto.upload;
 
 import lombok.Builder;
 import lombok.Getter;
